refactor: move ObjectDetail components into own folder in 'app'
