+++ src/app/catalog/_components/ObjectDetail/ObjectHero.module.css
