+++ src/app/catalog/_components/ObjectDetail/ObjectProperties/TabButton.tsx
