+++ src/app/catalog/_components/ObjectDetail/PropItem.tsx
