+++ src/app/catalog/complexes/[complex]/object/[object]/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import ObjectDetail from '@/app/catalog/_components/ObjectDetail'
+import ObjectDetail from '@/app/catalog/_components/ObjectDetail/ObjectDetail'
 import { getObject } from '@/globals/api'
 import { ObjectPage } from '@/types/Page'
