+++ src/app/catalog/new-building/object/[object]/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ObjectPage, SubcategoryPage } from '@/types/Page'
 import { getObject } from '@/globals/api'
-import ObjectDetail from '@/app/catalog/_components/ObjectDetail'
+import ObjectDetail from '@/app/catalog/_components/ObjectDetail/ObjectDetail'
 import SimilarObjects from '@/app/catalog/_components/SimilarObjects'
 import { getSimilarObjects } from '@/globals/api/methods/catalog-details/getSimilarObjects'
