+++ src/app/catalog/secondary-housing/[subcategory]/object/[object]/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ObjectPage, SubcategoryPage } from '@/types/Page'
 import { getObject } from '@/globals/api'
-import ObjectDetail from '@/app/catalog/_components/ObjectDetail'
+import ObjectDetail from '@/app/catalog/_components/ObjectDetail/ObjectDetail'
 
 async function Page({ params }: ObjectPage & SubcategoryPage) {
   const object = await getObject({
